* Changed root mult back to long from int, started getting wrong mults at high levels. * Fixed a small bug when adding connections in the dynkin diagram.

+++ trunk/src/edu/rug/hep/simplie/algebra/CRoot.java
@@ -42,7 +42,7 @@ public class CRoot implements Serializable
 	/** The norm of the root (the innerproduct with itself). */
 	public short norm;
 	/** The root multiplicity. */
-	public int	mult;
+	public long	mult;
 	/** Sum over multiplicities of roots that are fractionals of this one (used in Peterson's formula). */
 	public fraction	coMult;
